requirements.txt
pandas
feedparser
gunicorn
psycopg2-binary
sqlalchemy
uvicorn
asyncio
aiohttp
python-dotenv
fastapi